+++ ts-sdk/client/tests/orca.test.ts
@@ -18,8 +18,8 @@ describe("Tuna Position on Orca", () => {
     testMarket = await setupTestMarket(
       {
         addressLookupTable: DEFAULT_ADDRESS,
-        borrowLimitA: 0n,
-        borrowLimitB: 0n,
+        borrowLimitA: 100_000_000_000_000n,
+        borrowLimitB: 100_000_000_000_000n,
         disabled: false,
         liquidationFee: 10000, // 1%
         liquidationThreshold: 920000, // 92%