+++ ts-sdk/client/tests/tunaOracle.test.ts
@@ -27,8 +27,8 @@ describe("Tuna Oracle", () => {
   beforeEach(async () => {
     const marketArgs = {
       addressLookupTable: DEFAULT_ADDRESS,
-      borrowLimitA: 0n,
-      borrowLimitB: 0n,
+      borrowLimitA: 100_000_000_000_000n,
+      borrowLimitB: 100_000_000_000_000n,
       disabled: false,
       liquidationFee: 10000, // 1%
       liquidationThreshold: 920000, // 92%